Move imap logic into better folder structure
Should help us separate some of the logic in multiple files.
